+++ alura-tracker/src/router/index.ts
@@ -2,12 +2,16 @@ import { RouteRecordRaw, createRouter, createWebHashHistory } from "vue-router";
 import Tarefas from "../views/Tarefas.vue";
 import Projetos from "../views/Projetos.vue";
 import Formulario from "../views/Projetos/Formulario.vue";
+import Lista from "../views/Projetos/Lista.vue";
 
 const rotas: RouteRecordRaw[] = [
     {path: "/", name: "Tarefas", component: Tarefas},
-    {path: "/projetos", name: "Projetos", component: Projetos},
-    {path: "/projetos/criar", name: "Novo Projeto", component: Formulario},
-    {path: "/projetos/:id", name: "Editar projeto", component: Formulario, props: true},
+    {path: "/projetos", component: Projetos, children: [
+        {path: "", name: "Projetos", component: Lista},
+        {path: "criar", name: "Novo Projeto", component: Formulario},
+        {path: ":id", name: "Editar projeto", component: Formulario, props: true},
+    ]},
+
 ]
 
 const roteador = createRouter({
